+++ src/main/java/org/example/GA/BCRCD_CrossoverTaskUp.java
@@ -14,7 +14,7 @@ import static org.example.GA.GeneticAlgorithm.conflictCheck;
 
 public class BCRCD_CrossoverTaskUp implements Runnable {
     private final GeneticAlgorithm ga;
-    private final int identity;
+    private final long identity;
     private final float mutRate;
     private boolean cross;
     private final int r1;
@@ -24,7 +24,7 @@ public class BCRCD_CrossoverTaskUp implements Runnable {
     private final Patient[] allPatients;
     private final double[][] distanceMatrix;
 
-    public BCRCD_CrossoverTaskUp(GeneticAlgorithm ga, int identity, float mutRate, Chromosome p1, Chromosome p2, int r1, int r2, boolean cross, InstancesClass data) {
+    public BCRCD_CrossoverTaskUp(GeneticAlgorithm ga, long identity, float mutRate, Chromosome p1, Chromosome p2, int r1, int r2, boolean cross, InstancesClass data) {
         this.ga = ga;
         this.identity = identity;
         this.mutRate = mutRate;
