+++ src/main/java/org/example/GA/BCRC_CrossoverTask.java
@@ -14,14 +14,14 @@ import static org.example.GA.GeneticAlgorithm.conflictCheck;
 
 public class BCRC_CrossoverTask implements Runnable {
     private final GeneticAlgorithm ga;
-    private final int identity;
+    private final long identity;
     private final float mutRate;
     private final boolean cross;
     private final int r;
     private final Chromosome p1, p2;
     private final InstancesClass data;
 
-    public BCRC_CrossoverTask(GeneticAlgorithm ga, int identity, float mutRate, Chromosome p1, Chromosome p2, int r, boolean cross, InstancesClass data) {
+    public BCRC_CrossoverTask(GeneticAlgorithm ga, long identity, float mutRate, Chromosome p1, Chromosome p2, int r, boolean cross, InstancesClass data) {
         this.ga = ga;
         this.identity = identity;
         this.mutRate = mutRate;
